pos/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
PyMySQL==1.1.1
openpyxl==3.1.2
Werkzeug==3.0.3
python-dotenv==1.0.1
